shard aware statement cache
rails 4.2 caches the generated ast for simple find and find_by
queries. The problem is when the use_qualified_names setting is enabled,
this cached ast quotes the name of the shard in the query. This means if
the same find or find_by is run in a new shard, the query will
incorrectly execute in the shard in which the ast was first generated

This commit modifies that caching mechanism to store information on the
shard in which the query is being run on the cache key so that the same
find or find_by run in another shard results in a cache miss

This also bumps the version and locks the gemspec to rails 5.0.0.beta3
since rc1 has introduced breaking changes that are incompatible with
switchman

Change-Id: Ia0009f34dce0cb39cde129bc12f0e7301173b950
Reviewed-on: https://gerrit.instructure.com/80607
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Anthus Williams <awilliams@instructure.com>
QA-Review: Anthus Williams <awilliams@instructure.com>