use new dockerized jenkins jobs
This provides a Docker environment for use locally as well as on Jenkins
to run the entire test suite in a full build matrix of all Ruby and
Rails versions supported.

This also enables code coverage reports on all suite runs by default,
and starts archiving those automatically in Jenkins builds.

Change-Id: Ib6eb0c009d247ae8963ce8c684ed4a92be5d7a6f