Don't attempt to call the nonexistent "from_value".
This is a bug that was introduced while dropping Rails 4.2 support. See
https://gerrit.instructure.com/c/switchman/+/120030/3/lib/switchman/active_record/query_methods.rb#205.
Instead of unconditionally calling the Rails 5 code (from_clause) and dropping the
Rails 4.2 code (from_value), it unconditionally calls both. Calling
from_value leads to a missing method, which attempts to delegate to
arel, which leads to a recursive call (which leads to a stack overflow).

Change-Id: I546680a918d670f4e67d8a87aab232f624856390
Reviewed-on: https://gerrit.instructure.com/188535
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins