Run tests separately on Rails 5.2.2 and 5.2.3
The arel_columns method was rewritten in Rails 5.2.3. The primary effect of this
rewrite is that `pluck` and `select` are handled differently when
there's a `from` clause. See
https://github.com/rails/rails/commit/e9b1b8d2f0cd3899067174b433121344e87eb878
https://github.com/rails/rails/pull/35360/files

The tests currently don't cover this case, which is why everything
passes. I intend to push a separate patch or patches to add tests and
fix things, but first I'm laying the groundwork for testing on these
versions separately.

Change-Id: I0e732ef05c944e755e454862b6f127536f1f5e5d
Reviewed-on: https://gerrit.instructure.com/193214
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>