Fix AssociationQueryValue bugs
Despite it's name, primary_key is not really always the primary key of
the model and is can be the foreign key in some cases (depending on
which side of the relation you are on).  Also, upstream rails now calls
the correct method internally for AssociationQueryValue, but they do
not for PolymorphicArrayValue, so we still need to override that one.

Change-Id: I290e9df35be001a3a5008a34fa5babe13d842859
Reviewed-on: https://gerrit.instructure.com/c/switchman/+/321967
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jacob Burroughs <jburroughs@instructure.com>
Product-Review: Jacob Burroughs <jburroughs@instructure.com>